Add group state information saving and retrieval to and from a map
The main purpose of this change is to provide group state information without going to the database for better performance. This avoids saturating the database with requests for those clients who would use the group state info retrieval to monitor group state change in a small time interval say 1 second.