Wait until busy screen is not visible before making "a wait when clickable - click"
The above change's purpose is to fix an intermittent bug wherein an element is not clickable since the "gray cover" of the busy screen is still covering the screen when the click is issued.
Relying solely on checking if the element is clickable seems to be not enough.

In addition:

- optimize wait click by setting the wait driver's sleep time to 100ms