Fix broken tests (@author: Sharvari Barve)
The tests broke due to a change in browser node element composition after upgrading to jquery UI 1.12.0

Note: The original author of this fix is Sharvari Barve. Due to the purging of sensitive data in a previous commit her code had to be reapplied by the one who did the purge that is why the commit name is not hers.